services/neptune-django/requirements.txt
Django==2.1.2
djangorestframework
djangorestframework-jwt
cryptography
pyjwt
python-dotenv
